arch: move User-related code to :core-user
