+++ context-registry/build.gradle
@@ -44,6 +44,7 @@ android {
 dependencies {
     implementation project(':core-registry')
     implementation project(':core-session')
+    implementation project(':core-user')
     implementation project(':platform-design-system')
     implementation project(':platform-loadable')
     implementation libraries.coil.compose
