+++ context-registry/src/main/java/com/jeanbarrossilva/ongoing/context/registry/extensions/CollectionExtensions.kt
@@ -3,7 +3,7 @@ package com.jeanbarrossilva.ongoing.context.registry.extensions
 import com.jeanbarrossilva.ongoing.context.registry.domain.activity.ContextualActivity
 import com.jeanbarrossilva.ongoing.core.registry.activity.Activity
 import com.jeanbarrossilva.ongoing.core.session.SessionManager
-import com.jeanbarrossilva.ongoing.core.session.user.UserRepository
+import com.jeanbarrossilva.ongoing.core.user.UserRepository
 
 suspend fun Collection<Activity>.mapToContextualActivity(
     sessionManager: SessionManager,
