+++ core-user-in-memory/build.gradle
@@ -1,0 +1,13 @@
+plugins {
+    id 'java-library'
+    id 'org.jetbrains.kotlin.jvm'
+}
+
+java {
+    sourceCompatibility = versions.java
+    targetCompatibility = versions.java
+}
+
+dependencies {
+    implementation project(':core-user')
+}
