+++ core-user/src/main/java/com/jeanbarrossilva/ongoing/core/user/UserRepository.kt
@@ -1,4 +1,4 @@
-package com.jeanbarrossilva.ongoing.core.session.user
+package com.jeanbarrossilva.ongoing.core.user
 
 interface UserRepository {
     suspend fun getUserById(id: String): User?
