+++ feature-activities/build.gradle
@@ -43,6 +43,7 @@ android {
 dependencies {
     implementation project(':core-registry')
     implementation project(':core-session')
+    implementation project(':core-user')
     implementation project(':platform-design-system')
     implementation project(':platform-loadable')
     implementation project(':context-registry')
@@ -50,11 +51,14 @@ dependencies {
     implementation libraries.navigation.runtime
 
     testImplementation project(':core-session-in-memory')
+    testImplementation project(':core-user-in-memory')
     testImplementation libraries.kotlinx.coroutines.test
     testImplementation libraries.test.ext.junit
     testImplementation libraries.turbine
 
+    androidTestImplementation project(':core-user-in-memory')
     androidTestImplementation project(':platform-registry-test')
     androidTestImplementation libraries.compose.ui.test.junit4
     androidTestImplementation libraries.compose.ui.test.manifest
+    androidTestImplementation libraries.turbine
 }
