+++ feature-activity-details/build.gradle
@@ -44,6 +44,7 @@ dependencies {
     implementation project(':context-registry')
     implementation project(':core-registry')
     implementation project(':core-session')
+    implementation project(':core-user')
     implementation project(':platform-design-system')
     implementation project(':platform-extensions')
     implementation project(':platform-loadable')
