+++ feature-settings/build.gradle
@@ -44,12 +44,14 @@ dependencies {
     implementation project(':context-registry')
     implementation project(':core-registry')
     implementation project(':core-session')
+    implementation project(':core-user')
     implementation project(':platform-design-system')
     implementation project(':platform-extensions')
     implementation project(':platform-loadable')
     implementation libraries.compose_settings
 
     androidTestImplementation project(':core-session-in-memory')
+    androidTestImplementation project(':core-user-in-memory')
     androidTestImplementation project(':platform-registry-test')
     androidTestImplementation libraries.compose.ui.test.junit4
     androidTestImplementation libraries.compose.ui.test.manifest
