+++ platform-registry-test/src/main/java/com/jeanbarrossilva/ongoing/platform/registry/test/sessionmanager/SessionManagerRule.kt
@@ -1,4 +1,4 @@
-package com.jeanbarrossilva.ongoing.platform.registry.test.session
+package com.jeanbarrossilva.ongoing.platform.registry.test.sessionmanager
 
 import com.jeanbarrossilva.ongoing.core.session.Session
 import com.jeanbarrossilva.ongoing.core.session.SessionManager
