+++ settings.gradle
@@ -16,3 +16,5 @@ include ':feature-settings'
 include ':platform-loadable'
 include ':core-registry-in-memory'
 include ':core-extensions'
+include ':core-user'
+include ':core-user-in-memory'
